pyproject.toml
[project]
name = "notion_discord_sync"
version = "0.1.0"
description = "An integration tool between Notion and Discord"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "fastapi>=0.115.14",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]
